.claude/worktrees/naughty-solomon/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
openai==1.82.0
aiosqlite==0.21.0
python-dotenv==1.1.0
pydantic==2.11.3
websockets==15.0.1
